requirements.txt
mongo
bcrypt
requests
django
filetype
channels[daphne]
redis
channels-redis